switch to re-entrant methods
we switch from GEOSconnection (with a `status::Symbol`) to GEOScontext
(with a `ptr::GEOSContextHandle_t`), since the former is deprecated for
the latter. This change only applies to everything after GEOS v3.5.0.

There will be a global-level `_context` object that is passed to every
method to maintain the same behavior for now. We might want to also
provide methods that allows users to pass in their own context, and
fallback on the global _context object otherwise.