Switch to re-entrant methods (#46)
* switch to re-entrant methods

we switch from GEOSconnection (with a `status::Symbol`) to GEOScontext
(with a `ptr::GEOSContextHandle_t`), since the former is deprecated for
the latter. This change only applies to everything after GEOS v3.5.0.

There will be a global-level `_context` object that is passed to every
method to maintain the same behavior for now. We might want to also
provide methods that allows users to pass in their own context, and
fallback on the global _context object otherwise.

* expose `context::GEOScontext` as an optional argument to all methods

* switch to WK*Reader and WK*Writer style methods

- introduced `WKTReader`, `WKBReader`, `WKTWriter` and `WKBWriter` types
- we should be promoting the Julian (wrapper) geometry types through
`readgeom` and `writegeom`, rather than encouraging the C-pointers in
`_readgeom` and `_writegeom`
- the corresponding changes are `parseWKT` -> `readgeom`, `geomFromWKT`
-> `_readgeom`, and `geomToWKT` -> `_writegeom`
- moved the `GEOMTYPE` dict into `src/geos_operations.jl` since
`src/geos_functions.jl` is meant more for a direct 1-1 mapping with the
corresponding C functions

* bugfix

* remove module readers and writers in favor of ones local to methods

* update docs

* add deprecated functions

for a smoother transition

* GEOScontext -> GEOSContext

* test readgeom instead of _readgeom

* bugfix