rewrap using new Clang.jl
And include the used `wrap_geos.jl`, such that we can more easily wrap
it again in the future.

As far as I can tell this is breaking, but not too badly,
mainly because of some `const` that are now wrapped as a `@cenum`.
This also adds the CEnum dependency, Clang.jl uses this instead of the normal
julia `@enum` for C compatibility.