use GEOS_jll GEOS 3.8.0 artifacts
This will require at least julia 1.3. Testing to see if it works.
Some tests with pointOnSurface had to be updated. This is expected since it can return any point on the surface. Other than that some GeometryCollections are now ordered differently, and the unary union of an empty point remains an empty point.